front end gains

+++ TexasHoldEm.Library/Player.cs
@@ -14,8 +14,8 @@ namespace TexasHoldEm.Library
 
         public string Name { get; set; } = string.Empty;
         public int Position { get; set; } = 0;
-        public int Chips { get; set; } = 0;
-        public int CurrentBet { get; set; } = 0;
+        public double Chips { get; set; } = 0;
+        public  double CurrentBet { get; set; } = 0;
         public bool Folded { get; set; } = false;
         public bool AllIn { get; set; } = false;
         public Card[] Hand { get; } = new Card[2];
